- Issue 1733: keepgenerated not working with JDK6
  Reviewed by Jan Luehe.
  Passed QL
Issue number:
Obtained from:
Submitted by:
Reviewed by:


Original author: kchung
Date: 2006-12-14 17:48:35+00:00

svn path=/trunk/; revision=715

+++ src/java/org/apache/jasper/compiler/AntJavaCompiler.java
@@ -171,7 +171,7 @@ public class AntJavaCompiler implements JavaCompiler {
     }
 
     public void saveClassFile(String className, String classFileName) {
-        // class file are alwyas saved.
+        // class files are alwyas saved.
     }
 
     public void setDebug(boolean debug) {
@@ -285,9 +285,11 @@ public class AntJavaCompiler implements JavaCompiler {
         return javacErrors;
     }
 
-    public void removeJavaFile() {
-        File javaFile = new File(javaFileName);
-        javaFile.delete();
+    public void doJavaFile(boolean keep) {
+        if (!keep) {
+            File javaFile = new File(javaFileName);
+            javaFile.delete();
+        }
     }
 
     public static void startThreadPool() {
